simulation endpoints work
