+++ package.json
@@ -23,8 +23,10 @@
   "dependencies": {
     "axios": "^1.12.2",
     "cors": "^2.8.5",
+    "csvtojson": "^2.0.10",
     "dotenv": "^17.2.3",
     "express": "^5.1.0",
+    "nodemon": "^3.1.10",
     "three": "^0.180.0"
   }
 }
